fix: generate types from JSDoc
