+++ lib/http-outgoing.js
@@ -3,26 +3,84 @@
 import { PassThrough } from 'stream';
 import assert from 'assert';
 
+/**
+ * @typedef {object} PodiumClientHttpOutgoingOptions
+ * @property {string} name
+ * @property {string} uri To the podlet's `manifest.json`
+ * @property {number} timeout In milliseconds
+ * @property {number} maxAge
+ * @property {number} [retries=4]
+ * @property {boolean} [throwable=false]
+ * @property {boolean} [redirectable=false]
+ * @property {boolean} [rejectUnauthorized=true]
+ * @property {import('http').Agent} [httpAgent]
+ * @property {import('https').Agent} [httpsAgent]
+ */
+
+/**
+ * @typedef {object} PodiumClientResourceOptions
+ * @property {string} [pathname]
+ * @property {import('http').IncomingHttpHeaders} [headers]
+ * @property {object} [query]
+ */
+
+/**
+ * @typedef {object} PodiumRedirect
+ * @property {number} statusCode;
+ * @property {string} location;
+ */
+
+/**
+ * @typedef {object} PodletProxySchema
+ * @property {string} target
+ * @property {string} name
+ */
+
+/**
+ * @typedef {object} PodletManifest Similar to the schema's manifest, but with instances of AssetCss and AssetJs from `@podium/utils` and default values.
+ * @property {string} name
+ * @property {string} version
+ * @property {string} content
+ * @property {string} fallback
+ * @property {Array<import('@podium/utils').AssetJs>} js
+ * @property {Array<import('@podium/utils').AssetCss>} css
+ * @property {Record<string, string> | Array<PodletProxySchema>} proxy
+ * @property {string} team
+ */
+
 export default class PodletClientHttpOutgoing extends PassThrough {
     #rejectUnauthorized;
     #killRecursions;
     #killThreshold;
     #redirectable;
     #reqOptions;
-    #isFallback;
+    #isFallback = false;
     #throwable;
+    /** @type {PodletManifest} */
     #manifest;
     #incoming;
-    #redirect;
+    /** @type {null | PodiumRedirect} */
+    #redirect = null;
     #timeout;
     #success;
     #headers;
     #maxAge;
+    /** @type {'empty' | 'fresh' | 'cached' | 'stale'} */
     #status;
     #name;
-    #uri;    
-    constructor(
-        {
+    #uri;
+
+    /**
+     * @constructor
+     * @param {PodiumClientHttpOutgoingOptions} options
+     * @param {PodiumClientResourceOptions} [reqOptions]
+     * @param {import('@podium/utils').HttpIncoming} [incoming]
+     */
+    // @ts-expect-error Deliberate default empty options for better error messages
+    constructor(options = {}, reqOptions, incoming) {
+        super();
+
+        const {
             rejectUnauthorized = true,
             throwable = false,
             redirectable = false,
@@ -31,11 +89,7 @@ export default class PodletClientHttpOutgoing extends PassThrough {
             maxAge,
             name = '',
             uri,
-        } = {},
-        reqOptions,
-        incoming,
-    ) {
-        super();
+        } = options;
 
         assert(
             uri,
@@ -45,7 +99,6 @@ export default class PodletClientHttpOutgoing extends PassThrough {
         // If requests to https sites should reject on unsigned sertificates
         this.#rejectUnauthorized = rejectUnauthorized;
 
-        // A HttpIncoming object
         this.#incoming = incoming;
 
         // Kill switch for breaking the recursive promise chain
@@ -67,6 +120,7 @@ export default class PodletClientHttpOutgoing extends PassThrough {
         // Manifest which is either retrieved from the registry or
         // remote podlet (in other words its not saved in registry yet)
         this.#manifest = {
+            // @ts-expect-error Internal property
             _fallback: '',
         };
 
@@ -82,14 +136,6 @@ export default class PodletClientHttpOutgoing extends PassThrough {
         // How long the manifest should be cached before refetched
         this.#maxAge = maxAge;
 
-        // What status the manifest is in. This is used to tell what actions need to
-        // be performed throughout the resolving process to complete a request.
-        //
-        // The different statuses can be:
-        // "empty" - there is no manifest available - we are in process of fetching it
-        // "fresh" - the manifest has been fetched but is not stored in cache yet
-        // "cached" - the manifest was retrieved from cache
-        // "stale" - the manifest is outdated, a new manifest needs to be fetched
         this.#status = 'empty';
 
         // Name of the resource (name given to client)
@@ -100,13 +146,6 @@ export default class PodletClientHttpOutgoing extends PassThrough {
 
         // Whether the user can handle redirects manually.
         this.#redirectable = redirectable;
-
-        // When redirectable is true, this object should be populated with redirect information
-        // such that a user can perform manual redirection
-        this.#redirect = null;
-
-        // When isfallback is true, content fetch has failed and fallback will be served instead
-        this.#isFallback = false;
     }
 
     get rejectUnauthorized() {
@@ -130,10 +169,12 @@ export default class PodletClientHttpOutgoing extends PassThrough {
     }
 
     get fallback() {
+        // @ts-expect-error Internal property
         return this.#manifest._fallback;
     }
 
     set fallback(value) {
+        // @ts-expect-error Internal property
         this.#manifest._fallback = value;
     }
 
@@ -169,6 +210,16 @@ export default class PodletClientHttpOutgoing extends PassThrough {
         this.#maxAge = value;
     }
 
+    /**
+     * What status the manifest is in. This is used to tell what actions need to
+     * be performed throughout the resolving process to complete a request.
+     *
+     * The different statuses can be:
+     *  - `"empty"` - there is no manifest available - we are in process of fetching it
+     *  - `"fresh"` - the manifest has been fetched but is not stored in cache yet
+     *  - `"cached"` - the manifest was retrieved from cache
+     *  - `"stale"` - the manifest is outdated, a new manifest needs to be fetched
+     */
     get status() {
         return this.#status;
     }
@@ -193,18 +244,33 @@ export default class PodletClientHttpOutgoing extends PassThrough {
         return this.#manifest.content;
     }
 
+    /**
+     * Kill switch for breaking the recursive promise chain in case it is never able to completely resolve.
+     * This is true if the number of recursions matches the threshold.
+     */
     get kill() {
         return this.#killRecursions === this.#killThreshold;
     }
 
+    /**
+     * The number of recursions before the request should be {@link kill}ed
+     */
     get recursions() {
         return this.#killRecursions;
     }
 
+    /**
+     * Set the number of recursions before the request should be {@link kill}ed
+     */
     set recursions(value) {
         this.#killRecursions = value;
     }
 
+    /**
+     * When {@link redirectable} is `true` this is populated with redirect information so you can send a redirect response to the browser from your layout.
+     *
+     * @see https://podium-lib.io/docs/layout/handling_redirects
+     */
     get redirect() {
         return this.#redirect;
     }
@@ -213,6 +279,11 @@ export default class PodletClientHttpOutgoing extends PassThrough {
         this.#redirect = value;
     }
 
+    /**
+     * Whether the podlet can signal redirects to the layout.
+     *
+     * @see https://podium-lib.io/docs/layout/handling_redirects
+     */
     get redirectable() {
         return this.#redirectable;
     }
@@ -222,17 +293,22 @@ export default class PodletClientHttpOutgoing extends PassThrough {
     }
 
     /**
-     * Boolean getter that indicates whether the client is responding with a content or fallback payload.
+     * True if the client has returned the podlet's fallback.
+     *
      * @example
-     * ```
+     *
+     * ```js
      * if (outgoing.isFallback) console.log("Fallback!");
      * ```
+     *
+     * @see https://podium-lib.io/docs/podlet/fallbacks
      */
     get isFallback() {
-      return this.#isFallback;
+        return this.#isFallback;
     }
 
     pushFallback() {
+        // @ts-expect-error Internal property
         this.push(this.#manifest._fallback);
         this.push(null);
         this.#isFallback = true;
@@ -241,4 +317,4 @@ export default class PodletClientHttpOutgoing extends PassThrough {
     get [Symbol.toStringTag]() {
         return 'PodletClientHttpOutgoing';
     }
-};
+}
