+++ lib/resolver.cache.js
@@ -5,9 +5,20 @@ import clonedeep from 'lodash.clonedeep';
 import abslog from 'abslog';
 import assert from 'assert';
 
+/**
+ * @typedef {object} PodletClientCacheResolverOptions
+ * @property {import('abslog').AbstractLoggerOptions} [logger]
+ */
+
 export default class PodletClientCacheResolver {
     #registry;
     #log;
+
+    /**
+     * @constructor
+     * @param {import('ttl-mem-cache').default} registry
+     * @param {PodletClientCacheResolverOptions} options
+     */
     constructor(registry, options = {}) {
         assert(
             registry,
@@ -17,8 +28,14 @@ export default class PodletClientCacheResolver {
         this.#log = abslog(options.logger);
     }
 
+    /**
+     * Loads the podlet's manifest from cache if not stale
+     *
+     * @param {import('./http-outgoing.js').default} outgoing
+     * @returns {Promise<import('./http-outgoing.js').default>}
+     */
     load(outgoing) {
-        return new Promise(resolve => {
+        return new Promise((resolve) => {
             if (outgoing.status !== 'stale') {
                 const cached = this.#registry.get(outgoing.name);
                 if (cached) {
@@ -33,8 +50,14 @@ export default class PodletClientCacheResolver {
         });
     }
 
+    /**
+     * Saves the podlet's manifest to the cache
+     *
+     * @param {import('./http-outgoing.js').default} outgoing
+     * @returns {Promise<import('./http-outgoing.js').default>}
+     */
     save(outgoing) {
-        return new Promise(resolve => {
+        return new Promise((resolve) => {
             if (outgoing.status === 'fresh') {
                 this.#registry.set(
                     outgoing.name,
@@ -55,4 +78,4 @@ export default class PodletClientCacheResolver {
     get [Symbol.toStringTag]() {
         return 'PodletClientCacheResolver';
     }
-};
+}
