+++ lib/resolver.content.js
@@ -21,11 +21,24 @@ const pkg = JSON.parse(pkgJson);
 
 const UA_STRING = `${pkg.name} ${pkg.version}`;
 
+/**
+ * @typedef {object} PodletClientContentResolverOptions
+ * @property {string} clientName
+ * @property {import('./http.js').default} [http]
+ * @property {import('abslog').AbstractLoggerOptions} [logger]
+ */
+
 export default class PodletClientContentResolver {
     #log;
     #metrics;
     #histogram;
     #http;
+
+    /**
+     * @constructor
+     * @param {PodletClientContentResolverOptions} options
+     */
+    // @ts-expect-error Deliberate default empty options for better error messages
     constructor(options = {}) {
         this.#http = options.http || new HTTP();
         const name = options.clientName;
@@ -54,6 +67,12 @@ export default class PodletClientContentResolver {
         return this.#metrics;
     }
 
+    /**
+     * Resolves/fetches the podlet's content.
+     *
+     * @param {import('./http-outgoing.js').default} outgoing
+     * @returns {Promise<import('./http-outgoing.js').default>}
+     */
     async resolve(outgoing) {
         if (outgoing.kill && outgoing.throwable) {
             this.#log.warn(
@@ -71,12 +90,12 @@ export default class PodletClientContentResolver {
             outgoing.success = true;
             outgoing.pushFallback();
             outgoing.emit(
-              'beforeStream',
-              new Response({
-                  js: utils.filterAssets("fallback", outgoing.manifest.js),
-                  css: utils.filterAssets("fallback", outgoing.manifest.css),
-              }),
-          );
+                'beforeStream',
+                new Response({
+                    js: utils.filterAssets('fallback', outgoing.manifest.js),
+                    css: utils.filterAssets('fallback', outgoing.manifest.css),
+                }),
+            );
             return outgoing;
         }
 
@@ -94,12 +113,12 @@ export default class PodletClientContentResolver {
             outgoing.success = true;
             outgoing.pushFallback();
             outgoing.emit(
-              'beforeStream',
-              new Response({
-                  js: utils.filterAssets("fallback", outgoing.manifest.js),
-                  css: utils.filterAssets("fallback", outgoing.manifest.css),
-              }),
-          );
+                'beforeStream',
+                new Response({
+                    js: utils.filterAssets('fallback', outgoing.manifest.js),
+                    css: utils.filterAssets('fallback', outgoing.manifest.css),
+                }),
+            );
             return outgoing;
         }
 
@@ -113,8 +132,9 @@ export default class PodletClientContentResolver {
         const uri = putils.uriBuilder(
             outgoing.reqOptions.pathname,
             outgoing.contentUri,
-        )
+        );
 
+        /** @type {import('./http.js').PodiumHttpClientRequestOptions} */
         const reqOptions = {
             rejectUnauthorized: outgoing.rejectUnauthorized,
             bodyTimeout: outgoing.timeout,
@@ -183,11 +203,17 @@ export default class PodletClientContentResolver {
                 outgoing.success = true;
                 outgoing.pushFallback();
                 outgoing.emit(
-                  'beforeStream',
-                  new Response({
-                      js: utils.filterAssets("fallback", outgoing.manifest.js),
-                      css: utils.filterAssets("fallback", outgoing.manifest.css),
-                  }),
+                    'beforeStream',
+                    new Response({
+                        js: utils.filterAssets(
+                            'fallback',
+                            outgoing.manifest.js,
+                        ),
+                        css: utils.filterAssets(
+                            'fallback',
+                            outgoing.manifest.css,
+                        ),
+                    }),
                 );
 
                 // Body must be consumed; https://github.com/nodejs/undici/issues/583#issuecomment-855384858
@@ -227,6 +253,7 @@ export default class PodletClientContentResolver {
             if (outgoing.redirectable && statusCode >= 300) {
                 outgoing.redirect = {
                     statusCode,
+                    // @ts-expect-error TODO: look into what happens if the podlet returns more than one location header
                     location: hdrs && hdrs.location,
                 };
             }
@@ -235,8 +262,8 @@ export default class PodletClientContentResolver {
                 'beforeStream',
                 new Response({
                     headers: outgoing.headers,
-                    js: utils.filterAssets("content", outgoing.manifest.js),
-                    css: utils.filterAssets("content", outgoing.manifest.css),
+                    js: utils.filterAssets('content', outgoing.manifest.js),
+                    css: utils.filterAssets('content', outgoing.manifest.css),
                     redirect: outgoing.redirect,
                 }),
             );
@@ -260,10 +287,7 @@ export default class PodletClientContentResolver {
                 this.#log.warn(
                     `could not create network connection to remote resource when trying to request content - resource: ${outgoing.name} - url: ${uri}`,
                 );
-                throw badGateway(
-                    `Error reading content at ${uri}`,
-                    error,
-                );
+                throw badGateway(`Error reading content at ${uri}`, error);
             }
 
             timer({
@@ -280,12 +304,12 @@ export default class PodletClientContentResolver {
 
             outgoing.pushFallback();
             outgoing.emit(
-              'beforeStream',
-              new Response({
-                  js: utils.filterAssets("fallback", outgoing.manifest.js),
-                  css: utils.filterAssets("fallback", outgoing.manifest.css),
-              }),
-          );
+                'beforeStream',
+                new Response({
+                    js: utils.filterAssets('fallback', outgoing.manifest.js),
+                    css: utils.filterAssets('fallback', outgoing.manifest.css),
+                }),
+            );
 
             return outgoing;
         }
