+++ lib/resolver.js
@@ -7,12 +7,25 @@ import Fallback from './resolver.fallback.js';
 import Content from './resolver.content.js';
 import Cache from './resolver.cache.js';
 
+/**
+ * @typedef {object} PodletClientResolverOptions
+ * @property {string} clientName
+ * @property {import('abslog').AbstractLoggerOptions} [logger]
+ */
+
 export default class PodletClientResolver {
     #cache;
     #manifest;
     #fallback;
     #content;
     #metrics;
+
+    /**
+     * @constructor
+     * @param {import('ttl-mem-cache').default} registry
+     * @param {PodletClientResolverOptions} options
+     */
+    // @ts-expect-error Deliberate default empty options for better error messages
     constructor(registry, options = {}) {
         assert(
             registry,
@@ -22,12 +35,16 @@ export default class PodletClientResolver {
         const log = abslog(options.logger);
         const http = new HTTP();
         this.#cache = new Cache(registry, options);
-        this.#manifest = new Manifest({ ...options, http });
+        this.#manifest = new Manifest({
+            clientName: options.clientName,
+            logger: options.logger,
+            http,
+        });
         this.#fallback = new Fallback({ ...options, http });
         this.#content = new Content({ ...options, http });
         this.#metrics = new Metrics();
 
-        this.#metrics.on('error', error => {
+        this.#metrics.on('error', (error) => {
             log.error(
                 'Error emitted by metric stream in @podium/client module',
                 error,
@@ -43,14 +60,20 @@ export default class PodletClientResolver {
         return this.#metrics;
     }
 
+    /**
+     * Resolve the podlet's manifest, fallback and content
+     *
+     * @param {import('./http-outgoing.js').default} outgoing
+     * @returns {Promise<import('./http-outgoing.js').default>}
+     */
     resolve(outgoing) {
         return this.#cache
             .load(outgoing)
-            .then(obj => this.#manifest.resolve(obj))
-            .then(obj => this.#fallback.resolve(obj))
-            .then(obj => this.#content.resolve(obj))
-            .then(obj => this.#cache.save(obj))
-            .then(obj => {
+            .then((obj) => this.#manifest.resolve(obj))
+            .then((obj) => this.#fallback.resolve(obj))
+            .then((obj) => this.#content.resolve(obj))
+            .then((obj) => this.#cache.save(obj))
+            .then((obj) => {
                 if (obj.success) {
                     return obj;
                 }
@@ -58,15 +81,21 @@ export default class PodletClientResolver {
             });
     }
 
+    /**
+     * Refresh the podlet's cached manifest and fallback
+     *
+     * @param {import('./http-outgoing.js').default} outgoing
+     * @returns {Promise<boolean>} `true` if successful
+     */
     refresh(outgoing) {
         return this.#manifest
             .resolve(outgoing)
-            .then(obj => this.#fallback.resolve(obj))
-            .then(obj => this.#cache.save(obj))
-            .then(obj => !!obj.manifest.name);
+            .then((obj) => this.#fallback.resolve(obj))
+            .then((obj) => this.#cache.save(obj))
+            .then((obj) => !!obj.manifest.name);
     }
 
     get [Symbol.toStringTag]() {
         return 'PodletClientResolver';
     }
-};
+}
