+++ lib/resolver.manifest.js
@@ -20,11 +20,24 @@ const pkg = JSON.parse(pkgJson);
 
 const UA_STRING = `${pkg.name} ${pkg.version}`;
 
+/**
+ * @typedef {object} PodletClientManifestResolverOptions
+ * @property {string} clientName
+ * @property {import('abslog').AbstractLoggerOptions} [logger]
+ * @property {import('./http.js').default} [http]
+ */
+
 export default class PodletClientManifestResolver {
     #log;
     #metrics;
     #histogram;
     #http;
+
+    /**
+     * @constructor
+     * @param {PodletClientManifestResolverOptions} options
+     */
+    // @ts-expect-error Deliberate default empty options for better error messages
     constructor(options = {}) {
         this.#http = options.http || new HTTP();
         const name = options.clientName;
@@ -53,6 +66,10 @@ export default class PodletClientManifestResolver {
         return this.#metrics;
     }
 
+    /**
+     * @param {import('./http-outgoing.js').default} outgoing
+     * @returns {Promise<import('./http-outgoing.js').default>}
+     */
     async resolve(outgoing) {
         if (outgoing.status === 'cached') {
             return outgoing;
@@ -62,6 +79,7 @@ export default class PodletClientManifestResolver {
             'User-Agent': UA_STRING,
         };
 
+        /** @type {import('./http.js').PodiumHttpClientRequestOptions} */
         const reqOptions = {
             rejectUnauthorized: outgoing.rejectUnauthorized,
             timeout: outgoing.timeout,
@@ -81,11 +99,11 @@ export default class PodletClientManifestResolver {
         );
 
         try {
-            const {
-                statusCode,
-                headers: hdrs,
-                body,
-            } = await this.#http.request(outgoing.manifestUri, reqOptions);
+            const response = await this.#http.request(
+                outgoing.manifestUri,
+                reqOptions,
+            );
+            const { statusCode, headers: hdrs, body } = response;
 
             // Remote responds but with an http error code
             const resError = statusCode !== 200;
@@ -105,7 +123,11 @@ export default class PodletClientManifestResolver {
                 return outgoing;
             }
 
-            const manifest = validateManifest(await body.json());
+            const manifest = validateManifest(
+                /** @type {import("@podium/schemas").PodletManifestSchema} */ (
+                    await body.json()
+                ),
+            );
 
             // Manifest validation error
             if (manifest.error) {
@@ -158,6 +180,7 @@ export default class PodletClientManifestResolver {
             );
 
             // Construct css and js objects with absolute URIs
+            // @ts-expect-error We assign here what will end up as PodletManifest as defined in http-outgoing.js
             manifest.value.css = manifest.value.css.map((obj) => {
                 obj.value = utils.uriRelativeToAbsolute(
                     obj.value,
@@ -166,6 +189,7 @@ export default class PodletClientManifestResolver {
                 return new utils.AssetCss(obj);
             });
 
+            // @ts-expect-error We assign here what will end up as PodletManifest as defined in http-outgoing.js
             manifest.value.js = manifest.value.js.map((obj) => {
                 obj.value = utils.uriRelativeToAbsolute(
                     obj.value,
@@ -179,9 +203,9 @@ export default class PodletClientManifestResolver {
 
                 If .proxy is and Array, the podlet are of version 6 or newer. The Array is then an
                 Array of proxy Objects ({ name: 'foo', target: '/' }) which is the new structure
-                wanted from version 6 and onwards so leave this structure untouched. 
+                wanted from version 6 and onwards so leave this structure untouched.
 
-                If .proxy is an Object, the podlet are of version 5 or older where the key of the 
+                If .proxy is an Object, the podlet are of version 5 or older where the key of the
                 Object is the key of the target. If so, convert the structure to the new structure
                 consisting of an Array of proxy Objects.
 
@@ -191,12 +215,12 @@ export default class PodletClientManifestResolver {
             if (Array.isArray(manifest.value.proxy)) {
                 // Build absolute proxy URIs
                 manifest.value.proxy = manifest.value.proxy.map((item) => ({
-                        target: utils.uriRelativeToAbsolute(
-                            item.target,
-                            outgoing.manifestUri,
-                        ),
-                        name: item.name,
-                    }));
+                    target: utils.uriRelativeToAbsolute(
+                        item.target,
+                        outgoing.manifestUri,
+                    ),
+                    name: item.name,
+                }));
             } else {
                 const proxies = [];
                 // Build absolute proxy URIs
@@ -216,6 +240,7 @@ export default class PodletClientManifestResolver {
                 END: Proxy backwards compabillity check and handling
             */
 
+            // @ts-expect-error We map to AssetCss and AssetJs above
             outgoing.manifest = manifest.value;
             outgoing.status = 'fresh';
