+++ lib/response.js
@@ -1,11 +1,25 @@
 const inspect = Symbol.for('nodejs.util.inspect.custom');
 
+/**
+ * @typedef {object} PodiumClientResponseOptions
+ * @property {string} [content]
+ * @property {object} [headers]
+ * @property {Array<import('@podium/utils').AssetJs>} [js]
+ * @property {Array<import('@podium/utils').AssetCss>} [css]
+ * @property {import('./http-outgoing.js').PodiumRedirect | null} [redirect]
+ */
+
 export default class PodiumClientResponse {
     #redirect;
     #content;
     #headers;
     #css;
     #js;
+
+    /**
+     * @constructor
+     * @param {PodiumClientResponseOptions} options
+     */
     constructor({
         content = '',
         headers = {},
@@ -71,4 +85,4 @@ export default class PodiumClientResponse {
     get [Symbol.toStringTag]() {
         return 'PodiumClientResponse';
     }
-};
+}
