+++ lib/state.js
@@ -2,12 +2,25 @@ import EventEmitter from 'events';
 
 const inspect = Symbol.for('nodejs.util.inspect.custom');
 
+/**
+ * @typedef {object} PodiumClientStateOptions
+ * @property {number} [resolveThreshold=10000]
+ * @property {number} [resolveMax=240000]
+ */
+
 export default class PodiumClientState extends EventEmitter {
+    /** @type {NodeJS.Timeout | undefined} */
     #thresholdTimer;
     #threshold;
+    /** @type {NodeJS.Timeout | undefined} */
     #maxTimer;
+    /** @type {"instantiated" | "stable" | "initializing" | "unhealthy" | "unstable"} */
     #state;
     #max;
+
+    /**
+     * @param {PodiumClientStateOptions} [options]
+     */
     constructor({
         resolveThreshold = 10 * 1000,
         resolveMax = 4 * 60 * 1000,
@@ -113,4 +126,4 @@ export default class PodiumClientState extends EventEmitter {
     get [Symbol.toStringTag]() {
         return 'PodiumClientState';
     }
-};
+}
