+++ lib/utils.js
@@ -1,14 +1,12 @@
 /**
- * Checks if a header Oject has a header.
- * Will return true if the header exist and are not an empty
- * String or a String of whitespace.
+ * Checks if a header object has a header.
+ * Will return true if the header exist and is not an empty
+ * string or a string of whitespace.
  *
- * @param {Object} headers A headers object
- * @param {String} header A header value
- *
- * @returns {Boolean}
+ * @param {object} headers
+ * @param {string} header
+ * @returns {boolean}
  */
-
 export const isHeaderDefined = (headers, header) => {
     if (headers[header] === undefined || headers[header].trim() === '') {
         return false;
@@ -21,31 +19,24 @@ export const isHeaderDefined = (headers, header) => {
  * the changelog event object emitted from the internal
  * cache registry.
  *
- * @param {Object} item A changelog event object
- *
- * @returns {Boolean}
+ * @param {object} item A changelog event object
+ * @returns {boolean}
  */
-
-export const hasManifestChange = item => {
+export const hasManifestChange = (item) => {
     const oldVersion = item.oldVal ? item.oldVal.version : '';
     const newVersion = item.newVal ? item.newVal.version : '';
     return oldVersion !== newVersion;
 };
 
-
 /**
- * Check if a value is a HttpIncoming object or not. If not, it
+ * Check if a value is a Podium HttpIncoming object or not. If not, it
  * assume the incoming value is a context
  *
- * @param {Object} incoming A object
- *
- * @returns {HttpIncoming}
+ * @param {object} incoming
+ * @returns {boolean}
  */
-export const validateIncoming = (incoming = {}) => (Object.prototype.toString.call(incoming) === '[object PodiumHttpIncoming]');
-
-/**
- * @typedef {import("@podium/utils").AssetCss | import("@podium/utils").AssetJs} Asset
- */
+export const validateIncoming = (incoming = {}) =>
+    Object.prototype.toString.call(incoming) === '[object PodiumHttpIncoming]';
 
 /**
  * Filter assets array based on scope.
@@ -53,9 +44,10 @@ export const validateIncoming = (incoming = {}) => (Object.prototype.toString.ca
  * If scope property is set to "all", asset will be included.
  * If scope is set to "content" and asset scope property is set to "fallback", asset will not be included
  * If scope is set to "fallback" and asset scope property is set to "content", asset will not be included
+ * @template {import("@podium/utils").AssetCss | import("@podium/utils").AssetJs} T[]
  * @param {"content" | "fallback" | "all"} scope
- * @param {Asset[]} assets
- * @returns {Asset[]}
+ * @param {T[]} assets
+ * @returns {T[]}
  *
  * @example
  * ```
@@ -69,7 +61,13 @@ export const filterAssets = (scope, assets) => {
     // if undefined or null, passthrough
     if (!assets) return assets;
     // if a non array value is given, throw
-    if (!Array.isArray(assets)) throw new TypeError(`Asset definition must be of type array. Got ${typeof assets}`);
+    if (!Array.isArray(assets))
+        throw new TypeError(
+            `Asset definition must be of type array. Got ${typeof assets}`,
+        );
     // filter the array of asset definitions to matchin scope or anything with all. Treat no scope the same as "all" for backwards compatibility.
-    return assets.filter(asset => !asset.scope || asset.scope === scope || asset.scope === "all");
+    return assets.filter(
+        (asset) =>
+            !asset.scope || asset.scope === scope || asset.scope === 'all',
+    );
 };
